requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cleanup==5.0.0
djangorestframework==3.11.0
drf-form-bootstrap-4==0.5.2
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
